+++ src/components/Input/Input.type.ts
@@ -14,4 +14,6 @@ export type InputType = {
   required?: boolean;
   testing?: boolean;
   as?: void | WebTarget;
+  onChange: React.Dispatch<React.SetStateAction<string>>
+  [rest: string]: unknown;
 };